iot-social-message-consumer/target/requirements.txt
kafka-python==2.2.15
slack_sdk==3.37.0